Consolidated two windows into one and renamed to 'MainWindow'
