--- LeerCopyWPF/Views/BitmapWindow.xaml
@@ -1,12 +1,0 @@
-﻿<Window x:Class="LeerCopyWPF.Views.BitmapWindow" x:Name="BitmapWin"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:LeerCopyWPF.Views"
-        mc:Ignorable="d" 
-        Height="450" Width="800" 
-        ResizeMode="NoResize" ShowInTaskbar="False"
-        WindowState="Maximized" WindowStyle="None">
-    <Image x:Name="BitmapImg"/>
-</Window>
