+++ LeerCopyWPF/Views/MainWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Name="TransparentOverlayWin" x:Class="LeerCopyWPF.Overlay"
+﻿<Window x:Name="MainWin" x:Class="LeerCopyWPF.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -20,13 +20,11 @@
             <Setter Property="Margin" Value="0,10,10,0"/>
         </Style>
     </Window.Resources>
-    <!-- Set window transparency without affecting content transparency -->
-    <Window.Background>
-        <SolidColorBrush Color="#FFCBCBCB" Opacity="0.1"></SolidColorBrush>
-    </Window.Background>
     <!-- Window content -->
     <Grid>
-        <StackPanel x:Name="LabelPanel" HorizontalAlignment="Right">
+        <Image x:Name="ScreenImg" Panel.ZIndex="0"/>
+        <Grid x:Name="TransparentOverlay" Background="#FFCBCBCB" Opacity="0.1" Panel.ZIndex="1"/>
+        <StackPanel x:Name="LabelPanel" HorizontalAlignment="Right" Panel.ZIndex="2">
             <TextBlock x:Name="SaveLbl" Text="{x:Static resx:Resources.SaveLblStr}"/>
             <TextBlock x:Name="CopyLbl" Text="{x:Static resx:Resources.CopyLblStr}"/>
             <TextBlock x:Name="EditLbl" Text="{x:Static resx:Resources.EditLblStr}"/>
@@ -37,6 +35,6 @@
             <TextBlock x:Name="TipsLbl" Text="{x:Static resx:Resources.TipsLblStr}"/>
             <TextBlock x:Name="QuitLbl" Text="{x:Static resx:Resources.QuitLblStr}"/>
         </StackPanel>
-        <Image x:Name="SelectionImg" Visibility="Hidden"/>
+        <Image x:Name="SelectionImg" Visibility="Hidden" Panel.ZIndex="3"/>
     </Grid>
 </Window>
